+++ dask/async.py
@@ -100,8 +100,8 @@ import sys
 import traceback
 from operator import add
 from toolz import concat, partial
-from .core import (istask, flatten, reverse_dict, get_dependencies, ishashable,
-        inline)
+from .core import istask, flatten, reverse_dict, get_dependencies, ishashable
+from .optimize import inline
 from .utils import deepmap
 
 def inc(x):
